# refactor(tls): refactor SNI parsing