Drastically faster execution times (up to over 5000x in some cases).
Significant reductions in memory usage.
Fewer or eliminated memory allocations, which will likely reduce the load on the garbage collector.

before:
```
goos: darwin
goarch: arm64
pkg: github.com/AlexxIT/pnproxy/internal/tls
Benchmark_parseSNI/Valid_SNI-8         	   10000	    109540 ns/op	    5972 B/op	      54 allocs/op
Benchmark_parseSNI/Valid_SNI#01-8      	   10000	    119282 ns/op	    6020 B/op	      56 allocs/op
Benchmark_parseSNI/No_SNI-8            	  853980	      1392 ns/op	    2882 B/op	      19 allocs/op
PASS
ok  	github.com/AlexxIT/pnproxy/internal/tls	3.806s
```

after:
```
goos: darwin
goarch: arm64
pkg: github.com/AlexxIT/pnproxy/internal/tls
Benchmark_parseSNI/Valid_SNI-8         	62727058	        21.12 ns/op	       8 B/op	       1 allocs/op
Benchmark_parseSNI/Valid_SNI#01-8      	44897088	        26.07 ns/op	      16 B/op	       1 allocs/op
Benchmark_parseSNI/No_SNI-8            	451447947	         2.644 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/AlexxIT/pnproxy/internal/tls	4.235s
```